+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.6.2
 
   * Fix items not relevant to API appearing in permissions for a Section
+  * Fix started and started? methods for staged badges which have been completly done
 
 ## Version 0.6.1